FIRecognition/requirements.txt
numpy==1.19.5
keras==2.4.3
matplotlib==3.3.2
pillow==8.0.1
tensorflow==2.4.0
mtcnn==0.1.0
opencv-python==4.4.0.44
scikit-learn==0.24.2
keras-applications==1.0.8
keras-preprocessing==1.1.2
h5py==2.10.0
protobuf==3.19.0
git+https://github.com/rcmalli/keras-vggface.git
flask

Welcome to my github